requirements/requirements.txt
click
loguru
packaging
python-dateutil
requests
setuptools
typing-extensions
types-python-dateutil
types-requests
python3-tabulate
